+++ passerelle_imio_ia_aes/models.py
@@ -2452,27 +2452,66 @@ class ApimsAesConnector(BaseResource):
         #pattern=r"^$",
         display_category="Journées pédagogiques",
         parameters={
-            "school_implantation_id": {
-                "example_value": 2,
-                "description": "ID de l'implantation scolaire"
-                },
-            "child_id": {
-                "example_value": 518,
-                "description": "ID de l'enfant"
-                }
-            } 
+            "parent_id": {
+                "example_value": 279,
+                "description": "ID du parent"             
+            },
+            "grouped_by_date": {
+                "example_value": "False",
+                "description": "Faut-il regrouper les éléments par date ?"
+            }
+        }
     )
-    def list_pedagogicals_days(self, request, school_implantation_id, child_id):
-        url = f"{self.server_url}/{self.aes_instance}/pedagogical-days?school_implantation_id={school_implantation_id}"
+    def list_pedagogical_days(self, request, parent_id, grouped_by_date=False):
+        url = f"{self.server_url}/{self.aes_instance}/pedagogical-days?parent_id={parent_id}"
         response = self.session.get(url)
         response.raise_for_status()
         data = response.json()
-        
         for item in data.get("items", []):
-            if "date" in item:
-                item["text"] = self.format_date(item["date"])
-                item["disabled"] = child_id in item["child_ids"]
+                item["text"] = f"{item['child_lastname']} {item['child_firstname']}"
+                item["disabled"] = item["is_child_already_registered"]
+                item["id"] = f"{item['activity_id']}_{item['activity_date_id']}_{item['child_id']}"
+        logging.info(f"GROUPED_BY_DATE: {grouped_by_date}")
+        if grouped_by_date == "date":
+            items_dict = {}
+            for item in data.get("items"):
+                if items_dict.get(item["date"]) is None:
+                    items_dict[item["date"]] = []
+                items_dict[item["date"]].append(item)
+            data["items"] = items_dict
         return data
+
+
+    @endpoint(
+        name="pedagogical-days",
+        methods=["get"],
+        perm="can_access",
+        description="Lister les journées pédagogiques pour une commune",
+        example_pattern="per-dates",
+        pattern=r"^per-dates$",
+        display_category="Journées pédagogiques",
+        parameters={
+            "parent_id": {
+                "example_value": 279,
+                "description": "ID du parent"
+            },
+        }
+    )
+    def list_pedagogical_days_per_dates(self, request, parent_id):
+        url = f"{self.server_url}/{self.aes_instance}/pedagogical-days?parent_id={parent_id}"
+        response = self.session.get(url)
+        response.raise_for_status()
+        data = response.json()
+        items = {}
+        for item in data.get("items", []):
+            if items.get(item["date"]) is None:
+                items[item["date"]] = []
+            item["text"] = f"{item['child_lastname']} {item['child_firstname']}"
+            item["disabled"] = item["is_child_already_registered"]
+            item["id"] = f"{item['activity_id']}_{item['activity_date_id']}_{item['child_id']}"
+            items[item["date"]].append(item)
+        result = {"data": [{"id": k,"text": k,"registrations": v} for k,v in items.items()]}
+        return result
 
     @endpoint(
         name="pedagogical-days",
@@ -2497,9 +2536,11 @@ class ApimsAesConnector(BaseResource):
                 "child_id": int(post_data["child_id"]),
                 "parent_id": int(post_data["parent_id"]),
             })
-
+        
         payload = {"registrations": registrations}
+        logger.info(payload)
         response = self.session.post(url, json=payload)
         response.raise_for_status()
         return response.json()
+